Features/sfml

+++ Maize/src/Maize/Events/MouseEvent.h
@@ -35,31 +35,31 @@ namespace Maize {
 	class MouseMovedEvent : public Event
 	{
 	public:
-		MouseMovedEvent(const float x, const float y) : m_MouseX(x), m_MouseY(y) { }
+		MouseMovedEvent(const int32_t x, const int32_t y) : m_MouseX(x), m_MouseY(y) { }
 
-		float GetX() const { return m_MouseX; }
-		float GetY() const { return m_MouseY; }
+		int32_t GetX() const { return m_MouseX; }
+		int32_t GetY() const { return m_MouseY; }
 
 		EVENT_CLASS_TYPE(MouseMoved)
 
 	private:
-		float m_MouseX;
-		float m_MouseY;
+		int32_t m_MouseX;
+		int32_t m_MouseY;
 	};
 
 	class MouseScrolledEvent : public Event
 	{
 	public:
-		MouseScrolledEvent(const float xOffset, const float yOffset) : m_XOffset(xOffset), m_YOffset(yOffset) {}
+		MouseScrolledEvent(const int32_t xOffset, const int32_t yOffset) : m_XOffset(xOffset), m_YOffset(yOffset) {}
 
-		float GetXOffset() const { return m_XOffset; }
-		float GetYOffset() const { return m_YOffset; }
+		int32_t GetXOffset() const { return m_XOffset; }
+		int32_t GetYOffset() const { return m_YOffset; }
 
 		EVENT_CLASS_TYPE(MouseScrolled)
 
 	private:
-		float m_XOffset;
-		float m_YOffset;
+		int32_t m_XOffset;
+		int32_t m_YOffset;
 	};
 
 }
